Now Fraction has an int cast, so we don't need int(double( . . .

+++ ptclassic/src/domains/mdsdf/kernel/MDSDFStar.cc
@@ -121,8 +121,8 @@ int MDSDFStar::notRunnable() {
 
   // Check to see whether the required number of row and column repetitions
   //   has been met
-  if(int(double(rowRepetitions)) <= rowIndex ||
-     int(double(colRepetitions)) <= colIndex)
+  if(int(rowRepetitions) <= rowIndex ||
+     int(colRepetitions) <= colIndex)
     return 2;
 
   // Step through all the input ports, checking to see whether the required
